# ipx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Solana transaction API",
        "title": "SVMlab API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:33153",
    "basePath": "/",
    "paths": {
...
        "/svm/rpc/": {
            "post": {
                "description": "Passes a method straight through, so anything this API has not wrapped stays reachable.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc"
                ],
                "summary": "Call any JSON-RPC method",
                "parameters": [
                    {
                        "description": "Method and params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.RawRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/rpc/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc"
                ],
                "summary": "Call several JSON-RPC methods in one round trip",
                "parameters": [
                    {
                        "description": "Calls",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.BatchRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/svm/token/account": {
            "post": {
                "description": "Parses the 165 bytes a token account holds. owner is not the runtime owner: that is the token program, which is what may write the data, while owner here is the wallet whose signature the program accepts. amount is base units with no scale of its own, so the mint is read as well to report amount_ui; when the mint cannot be read the base units are still returned and decimals is omitted rather than guessed. Works for both classic Token and Token-2022, whose extensions sit past the base layout.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Read an SPL Token holder account",
                "parameters": [
                    {
                        "description": "Token account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/token.AccountRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/token.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/token/mint": {
            "post": {
                "description": "Parses the 82 bytes a mint account holds. A mint is the closest thing Solana has to an ERC-20 contract, and what it lacks is the point: no balances and no allowances, since those live in separate token accounts owned by each holder. mint_authority and freeze_authority are omitted when absent, which is different from being the zero address; a mint whose mint authority was removed has a permanently fixed supply, and one initialized without a freeze authority can never gain one. Works for both classic Token and Token-2022, whose extensions sit past the base layout.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Read an SPL Token mint",
                "parameters": [
                    {
                        "description": "Mint account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/token.MintRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/token.MintResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "misc.BatchCall": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "params": {}
            }
        },
        "misc.BatchRequest": {
            "type": "object",
            "properties": {
                "calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/misc.BatchCall"
                    }
                }
            }
        },
...
        "misc.RawRequest": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string",
                    "example": "getEpochInfo"
                },
                "params": {}
            }
        },
...
        "token.AccountRequest": {
            "type": "object",
            "properties": {
                "token_account": {
                    "type": "string",
                    "example": "4qRgcVrSqs43Jy9n8w7H5EJh8FnRiTFnABjE37Dd3Ae5"
                }
            }
        },
        "token.AccountResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "amount_ui": {
                    "type": "string"
                },
                "close_authority": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "delegate": {
                    "type": "string"
                },
                "delegated_amount": {
                    "type": "string"
                },
                "exists": {
                    "type": "boolean"
                },
                "frozen": {
                    "type": "boolean"
                },
                "mint": {
                    "type": "string"
                },
                "native": {
                    "type": "boolean"
                },
                "owner": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "rent_reserve": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "token.MintRequest": {
            "type": "object",
            "properties": {
                "mint_account": {
                    "type": "string",
                    "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
                }
            }
        },
        "token.MintResponse": {
            "type": "object",
            "properties": {
                "decimals": {
                    "type": "integer"
                },
                "exists": {
                    "type": "boolean"
                },
                "freezable": {
                    "type": "boolean"
                },
                "freeze_authority": {
                    "type": "string"
                },
                "initialized": {
                    "type": "boolean"
                },
                "mint_authority": {
                    "type": "string"
                },
                "mintable": {
                    "type": "boolean"
                },
                "program": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "supply": {
                    "type": "string"
                },
                "supply_ui": {
                    "type": "string"
                }
            }
        },
...
    }
}